Feature/force shift end (#2)
* add force end shift

* fix inactive chat removal

* add get shift endpoint

* fix bugs

+++ ChatAssignmentApp.Core/Shifts/Commands/EndShiftCommand.cs
@@ -7,13 +7,17 @@ namespace ChatAssignmentApp.Core.Shifts.Commands
 {
     public class EndShiftCommand : IEndShiftCommand
     {
+        private readonly Configuration _config;
+
         private readonly IQueueService _queueService;
         private readonly IShiftStorageService _shiftStorageService;
 
         public EndShiftCommand(
+            Configuration configuration,
             IQueueService queueService,
             IShiftStorageService shiftStorageService)
         {
+            _config = configuration;
             _queueService = queueService;
             _shiftStorageService = shiftStorageService;
         }
@@ -26,8 +30,19 @@ namespace ChatAssignmentApp.Core.Shifts.Commands
             if (currentShift == null)
                 return new CommandResult<bool>(false, "Current shift not found. ");
 
-            await _queueService.DeleteQueues(currentShift.IsOverflowAgentsAvailable);
+            var mainQueueItemCount = await _queueService.GetQueueItemCount(
+                _config.RabbitMQConfiguration.MainChatQueueName);
 
+            var overflowQueueItemCount = await _queueService.GetQueueItemCount(
+                _config.RabbitMQConfiguration.OverflowChatQueueName);
+
+            if (mainQueueItemCount > 0 && overflowQueueItemCount > 0)
+            {
+                currentShift.ForceShiftEnd();
+                return new CommandResult<bool>(true);
+            }
+
+            await _queueService.DeleteQueues(currentShift.IsOverflowAgentsAvailable);
             _shiftStorageService.DeleteShift(currentShift);
 
             return new CommandResult<bool>(true);
